Fixed issue with integer comparisons

+++ jgrapht-opt/src/main/java/org/jgrapht/opt/graph/sparse/specifics/IncidenceMatrixSparseUndirectedSpecifics.java
@@ -169,7 +169,9 @@ public class IncidenceMatrixSparseUndirectedSpecifics
             int v = getEdgeSource(eId);
             int u = getEdgeTarget(eId);
 
-            if (v == sourceVertex && u == targetVertex || v == targetVertex && u == sourceVertex) {
+            if (v == sourceVertex.intValue() && u == targetVertex.intValue()
+                || v == targetVertex.intValue() && u == sourceVertex.intValue())
+            {
                 return eId;
             }
         }
@@ -199,7 +201,9 @@ public class IncidenceMatrixSparseUndirectedSpecifics
             int v = getEdgeSource(eId);
             int u = getEdgeTarget(eId);
 
-            if (v == sourceVertex && u == targetVertex || v == targetVertex && u == sourceVertex) {
+            if (v == sourceVertex.intValue() && u == targetVertex.intValue()
+                || v == targetVertex.intValue() && u == sourceVertex.intValue())
+            {
                 result.add(eId);
             }
         }
